changes on api controller structure
